store manager can see appropriate functionality from store show page
